+++ Watchit/src/DataBase/CastMembersData.java
@@ -44,7 +44,7 @@ public class CastMembersData implements Data<CastMember>,ReadableClass{
                 String gender = data[0];
                 String nationality = data[1];
                 String socialMediaLink = data[2];
-                Long size = Long.parseLong(data[3]);
+                Long size = (Long) Long.parseLong(data[3]);
                 List<String>contents=new ArrayList<>();
                 for (int i = 0;i<size;i++){
                     contents.add(data[4+i]);
@@ -189,8 +189,8 @@ public class CastMembersData implements Data<CastMember>,ReadableClass{
     }
 
     public void removeData(long Id){
-        while(getDataById(Id)!=null) {
-            CastMembers.remove(getDataById(Id));
+        while(getDataById(Long.valueOf(Id))!=null) {
+            CastMembers.remove(getDataById(Long.valueOf(Id)));
         }
     }
 }
